refactor: location 위치 변경
