+++ src/domain/models/location.ts
@@ -3,3 +3,10 @@ export interface AppLocation {
   y: number;
   level?: number;
 }
+export type TPlace = kakao.maps.services.PlacesSearchResultItem;
+export interface ILocation {
+  lat: number;
+  lng: number;
+  level: number;
+  place: TPlace | null;
+}
