+++ src/store/locationStore.ts
@@ -1,12 +1,5 @@
+import { ILocation } from "@/domain/models/location";
 import { create } from "zustand";
-
-export type TPlace = kakao.maps.services.PlacesSearchResultItem;
-export interface ILocation {
-  lat: number;
-  lng: number;
-  level: number;
-  place: TPlace | null;
-}
 
 export interface ILocationStatus {
   loading: boolean;
